feat: connect distributed rewards stat to subgraph data

+++ .graphclient/index.ts
@@ -283,7 +283,7 @@ export type LToken = {
   aprUpdates?: Maybe<Array<APRUpdate>>;
   activities?: Maybe<Array<Activity>>;
   rewardsMints?: Maybe<Array<RewardsMint>>;
-  totalMintedRewards: Scalars['BigDecimal'];
+  totalMintedRewards: Scalars['BigInt'];
 };
 
 
@@ -363,14 +363,14 @@ export type LToken_filter = {
   aprUpdates_?: InputMaybe<APRUpdate_filter>;
   activities_?: InputMaybe<Activity_filter>;
   rewardsMints_?: InputMaybe<RewardsMint_filter>;
-  totalMintedRewards?: InputMaybe<Scalars['BigDecimal']>;
-  totalMintedRewards_not?: InputMaybe<Scalars['BigDecimal']>;
-  totalMintedRewards_gt?: InputMaybe<Scalars['BigDecimal']>;
-  totalMintedRewards_lt?: InputMaybe<Scalars['BigDecimal']>;
-  totalMintedRewards_gte?: InputMaybe<Scalars['BigDecimal']>;
-  totalMintedRewards_lte?: InputMaybe<Scalars['BigDecimal']>;
-  totalMintedRewards_in?: InputMaybe<Array<Scalars['BigDecimal']>>;
-  totalMintedRewards_not_in?: InputMaybe<Array<Scalars['BigDecimal']>>;
+  totalMintedRewards?: InputMaybe<Scalars['BigInt']>;
+  totalMintedRewards_not?: InputMaybe<Scalars['BigInt']>;
+  totalMintedRewards_gt?: InputMaybe<Scalars['BigInt']>;
+  totalMintedRewards_lt?: InputMaybe<Scalars['BigInt']>;
+  totalMintedRewards_gte?: InputMaybe<Scalars['BigInt']>;
+  totalMintedRewards_lte?: InputMaybe<Scalars['BigInt']>;
+  totalMintedRewards_in?: InputMaybe<Array<Scalars['BigInt']>>;
+  totalMintedRewards_not_in?: InputMaybe<Array<Scalars['BigInt']>>;
   /** Filter for the block changed event. */
   _change_block?: InputMaybe<BlockChangedFilter>;
   and?: InputMaybe<Array<InputMaybe<LToken_filter>>>;
@@ -1181,7 +1181,7 @@ export type LTokenResolvers<ContextType = MeshContext, ParentType extends Resolv
   aprUpdates?: Resolver<Maybe<Array<ResolversTypes['APRUpdate']>>, ParentType, ContextType, RequireFields<LTokenaprUpdatesArgs, 'skip' | 'first'>>;
   activities?: Resolver<Maybe<Array<ResolversTypes['Activity']>>, ParentType, ContextType, RequireFields<LTokenactivitiesArgs, 'skip' | 'first'>>;
   rewardsMints?: Resolver<Maybe<Array<ResolversTypes['RewardsMint']>>, ParentType, ContextType, RequireFields<LTokenrewardsMintsArgs, 'skip' | 'first'>>;
-  totalMintedRewards?: Resolver<ResolversTypes['BigDecimal'], ParentType, ContextType>;
+  totalMintedRewards?: Resolver<ResolversTypes['BigInt'], ParentType, ContextType>;
   __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
 }>;
